Superadmin
Added a superadmin in our enum role and created a superadmin menu. I
chose the switch case for that output. What I want to is to make sure
that the correct admin gets the permission to add a location. When the
correct admin gets the permission a new menu choice will be avalable for that same admin to add a new location.