Case 5 in program.cs
I have created a loop that searches through all the users and the output
will be only admins with index. I've created a new list for only admins and I did it that way because I wanted it to be easier for the superadmin to choose the correct admin by number. When
the number is chosen the superadmin is able to assign a region for the
chosen admin.