Problem with case 5
I understood that in the end of case 5 the superadmin shouldnt just
write a region. The system should loop through all locations and connect
admin with a region from the location. Chris and I are looking for a
solution because we dont know yet how we want to solve it.