The case for admins to be assigned to a certain region 'long version'
So first of all I created a loop and maked sure that the system gives a
list of all admins by index. The user chooses a certain admin. After
that I used arrays for my regions by enums so I get a list of every
region. After that the user chooses a region for that certain admin.
When this was done I created a new menu choice for an admin to see where
the admin has been assigned.